ZKP-69 implements tests for asset pallet and removes some functions

+++ pallets/support/src/traits.rs
@@ -1,3 +1,3 @@
 pub trait AssetInterface {
-    fn get_default_collateral() -> u8;
+    fn get_default_collateral() -> u64;
 }
